Change the denominator to penalize having lots of vague matches
I was unhappy with having to use the number of data points as 'n', even
if some of these data points are low-weight conceptual matches. The
result was causing terms with 1 exact match and lots of vague conceptual
matches to look like significant score drivers.

Yesterday's math fixes mean that we can use the total weight of the
examples, not the number of examples, as 'n', and it will still be
greater than 1 as it needs to be.