project_migration cleanup, step 1: modernization of code in main()
Mostly renaming "account" to "workspace" in variables, and some light
reformatting.

Also fixing a bug: previously the script got all projects and pared it down to
just those whose "account_id" was the ID of the workspace being copied from.
We no longer return "account_id"; and in any case there's a better way to get
the projects of just one workspace, so we use that.  (We also only get the
project record fields we need.)