Basic PEP8 cleanups to subset_key_terms.py
Mostly line-length things, also renaming "skt" variable so as not to shadow
the "skt()" function; renaming some one-letter variables; etc.

Also, there was a dictionary assignment happening in a for-loop over every
item in a list (even though it was always assigning the same dictionary to
the same variable), so that had to stop.