Fix subset_key_terms() output; simplify doc-text logic
This changes the output of subset_key_terms(), which is called in a number of
places, but as far as I can tell, every time it's called, all that happens to
it is that it gets fed to create_skt_table(), so as long as that's prepared for
triples, no one else should care what's in the list.

It also changes the logic around getting the doc text when there's fewer than
three documents.  This is a little quirky, but so was the previous version, and
this is way more compact.