Remove dead (i.e., v4-using) code
Most of this code won't have worked for something like a year, so it's a good
bet that no one has tried to run it, i.e., no one has needed it, i.e., we don't
need it.  If we do need it, we can recreate it properly (perhaps with reference
to the git history).