Fix and rename get_as()
np.dot() returns an numpy float regardless of whether its arguments are a
vectors of numpy floats or lists of regular floats.  So converting every float
in the vector wasn't doing anything (other than slowing things down).  If the
intent was for the *output* to be a float, it wasn't doing that either; to be
safe, the function now does cast its return value to float().

Also, every time I saw "get_as", I thought, get as what?  Get as an integer,
or get as text, or...?  Renamed it to get_assoc(), which is much clearer about
what it does.