Move get_driver_fields onto ScoreDrivers object
In the two places that get_driver_fields is called, we now have to instantiate
a ScoreDrivers object, but that's good, because we'll be using the object more
and more in those places.

The method itself is now somewhat simplified: it doesn't require a client,
because the ScoreDrivers object already has a client; and it doesn't have to
make an API call itself, but can rather use self.metadata, which may have
cached information from a previous call to get metadata.