Move get_first_date_field() onto ScoreDrivers
Behold, the refactoring begins to pay off!  Previous callers of this function
were also calling get_driver_fields().  Now that both of them are on a single
object, they can share the result of a single API call to get the metadata,
rather than each having to make an API call.

("Get metadata" is probably a pretty fast API call, so it's not a whole lot
of savings, but when you're talking about hitting an API, every little bit can
count.)