Consolidate write_table_to_csv
This code was roughly duplicated in score_drivers and bi_export_tool.  This
removes the "calc_keys" parameter in favor of always calculating keys (it's a
pretty quick dict comprehension, harmless when all of the dicts have the same
keys, necessary when they don't), and lets the export tool import it from score
drivers.